==> 3_PackageUsage/3_1_breastCancer/01_get_dataset.R <==
# install needed packages:
# install.packages("BiocManager")
# BiocManager::install("GEOquery")

gset <- GEOquery::getGEO("GSE3494", GSEMatrix = TRUE, AnnotGPL = TRUE)[[1]]

Biobase::fvarLabels(gset) <- make.names(Biobase::fvarLabels(gset))

gens <- gset@assayData$exprs

# Get the subset of data to analyze; 150 gens of 58 people:
data2 <- t(gens[
  c("1053_at", "200039_s_at", "200053_at", "200079_s_at", "200628_s_at", "200639_s_at", "200670_at", "200687_s_at", "200710_at", "200740_s_at", "200773_x_at", "200783_s_at", "200795_at", "200810_s_at", "200811_at", "200822_x_at", "200840_at", "200853_at", "200854_at", "200855_at", "200996_at", "201041_s_at", "201077_s_at", "201088_at", "201104_x_at", "201114_x_at", "201115_at", "201124_at", "201140_s_at", "201170_s_at", "201195_s_at", "201196_s_at", "201197_at", "201201_at", "201202_at", "201236_s_at", "201263_at", "201266_at", "201281_at", "201291_s_at", "201292_at", "201327_s_at", "201342_at", "201384_s_at", "201394_s_at", "201395_at", "201397_at", "201455_s_at", "201479_at", "201508_at", "201557_at", "201584_s_at", "201591_s_at", "201663_s_at", "201664_at", "201685_s_at", "201694_s_at", "201710_at", "201755_at", "201761_at", "201770_at", "201790_s_at", "201791_s_at", "201795_at", "201833_at", "201890_at", "201896_s_at", "201897_s_at", "201930_at", "201970_s_at", "201977_s_at", "201978_s_at", "202011_at", "202088_at", "202095_s_at", "202105_at", "202107_s_at", "202109_at", "202117_at", "202174_s_at", "202188_at", "202200_s_at", "202233_s_at", "202240_at", "202270_at", "202276_at", "202307_s_at", "202338_at", "202370_s_at", "202371_at", "202409_at", "202487_s_at", "202503_s_at", "202580_x_at", "202589_at", "202590_s_at", "202613_at", "202666_s_at", "202690_s_at", "202705_at", "202725_at", "202754_at", "202779_s_at", "202808_at", "202815_s_at", "202854_at", "202858_at", "202870_s_at", "202954_at", "202962_at", "203022_at", "203046_s_at", "203066_at", "203071_at", "203126_at", "203130_s_at", "203139_at", "203145_at", "203187_at", "203188_at", "203208_s_at", "203213_at", "203214_x_at", "203223_at", "203249_at", "203265_s_at", "203266_s_at", "203276_at", "203287_at", "203296_s_at", "203345_s_at", "203347_s_at", "203358_s_at", "203362_s_at", "203380_x_at", "203405_at", "203418_at", "203422_at", "203428_s_at", "203432_at", "203438_at", "203439_s_at", "203463_s_at", "203484_at", "203554_x_at", "203560_at", "203594_at", "203625_x_at", "203693_s_at", "203696_s_at"),
  c("GSM79115", "GSM79116", "GSM79118", "GSM79122", "GSM79123", "GSM79131", "GSM79138", "GSM79144", "GSM79145", "GSM79147", "GSM79148", "GSM79154", "GSM79159", "GSM79165", "GSM79188", "GSM79189", "GSM79191", "GSM79196", "GSM79202", "GSM79211", "GSM79216", "GSM79223", "GSM79227", "GSM79241", "GSM79242", "GSM79244", "GSM79247", "GSM79253", "GSM79259", "GSM79263", "GSM79270", "GSM79271", "GSM79280", "GSM79285", "GSM79287", "GSM79292", "GSM79298", "GSM79299", "GSM79302", "GSM79304", "GSM79305", "GSM79306", "GSM79313", "GSM79321", "GSM79325", "GSM79329", "GSM79333", "GSM79336", "GSM79338", "GSM79340", "GSM79342", "GSM79346", "GSM79349", "GSM79351", "GSM79352", "GSM79353", "GSM79356", "GSM79363")
])

save(data2, file = file.path(".", "3_PackageUsage", "3_1_breastCancer", "data", "data2.rda"))
